Major updates to allow for parallel building. Turn on PARALLEL=1 in cronjob.weekly and all build types for 1 arch will be run at once.
If you want to run more than that (i.e. multiple arches in parallel
too), then you'll need to tweak slightly more yet. Main issue is the
svn update and popcon update code, then the build lock file. More to
come on that front later...

Rather than rely on the exit status of testingcds, it will now write a
trace file as it exits. This will give us the end time and an error
code that we can parse safely.

Initial testing with 6 arches at once on the new CD build machine
shows that we can build those 6 arches in parallel in well under an
hour. The Squeeze release should be much quicker than Lenny was!